+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## v3.0.1 (2022-01-28)
 
 - Removed budge from `README.md` file that shows stable release;
+- Added `Option::JUST_NOW` option that print `Just now` if time is within 1 minute;
 
 ----